+++ addons/io_scene_swbf_msh/__init__.py
@@ -56,11 +56,12 @@ import bpy
 from bpy_extras.io_utils import ExportHelper, ImportHelper
 from bpy.props import BoolProperty, EnumProperty, CollectionProperty
 from bpy.types import Operator
-from .msh_scene import create_scene
+from .msh_scene_utilities import create_scene
 from .msh_scene_save import save_scene
 from .msh_scene_read import read_scene
 from .msh_material_properties import *
 from .msh_skeleton_properties import *
+from .msh_collision_prim_properties import *
 from .msh_to_blend import *
 from .zaa_to_blend import *
 
@@ -189,6 +190,8 @@ def menu_func_import(self, context):
 
 
 def register():
+    bpy.utils.register_class(CollisionPrimitiveProperties)
+
     bpy.utils.register_class(MaterialProperties)
     bpy.utils.register_class(MaterialPropertiesPanel)
 
@@ -201,11 +204,15 @@ def register():
     bpy.types.TOPBAR_MT_file_export.append(menu_func_export)
     bpy.types.TOPBAR_MT_file_import.append(menu_func_import)
 
-    bpy.types.Material.swbf_msh = bpy.props.PointerProperty(type=MaterialProperties)
+    bpy.types.Object.swbf_msh_coll_prim = bpy.props.PointerProperty(type=CollisionPrimitiveProperties)
+    bpy.types.Material.swbf_msh_mat = bpy.props.PointerProperty(type=MaterialProperties)
     bpy.types.Armature.swbf_msh_skel = bpy.props.CollectionProperty(type=SkeletonProperties)
 
 
+
 def unregister():
+    bpy.utils.unregister_class(CollisionPrimitiveProperties)
+
     bpy.utils.unregister_class(MaterialProperties)
     bpy.utils.unregister_class(MaterialPropertiesPanel)
