+++ addons/io_scene_swbf_msh/msh_material_to_blend.py
@@ -1,4 +1,4 @@
-""" For finding textures and assigning material properties from entries in a Material """
+""" For finding textures and assigning MaterialProperties from entries in a Material """
 
 import bpy
 from typing import Dict
